requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
matplotlib
scipy
scikit-learn
torchsummary
PyYAML
tqdm
mlflow
fastapi
uvicorn
python-multipart
prometheus-client
pytest
pytest-cov
httpx  # For FastAPI TestClient
requests
boto3
